Move Resolve* and NewCharmStore* functions to new store package.
