+++ cmd/juju/application/bundle/bundle.go
@@ -9,13 +9,11 @@ import (
 
 	"github.com/juju/bundlechanges"
 	"github.com/juju/charm/v7"
-	csparams "github.com/juju/charmrepo/v5/csclient/params"
 	"github.com/juju/errors"
 	"github.com/juju/loggo"
 	"github.com/juju/names/v4"
 
 	"github.com/juju/juju/apiserver/params"
-	"github.com/juju/juju/cmd/juju/application/utils"
 	"github.com/juju/juju/core/constraints"
 	"github.com/juju/juju/core/devices"
 	"github.com/juju/juju/core/model"
@@ -261,68 +259,4 @@ func verifyBundle(data *charm.BundleData, bundleDir string) error {
 		return errors.New("the provided bundle has the following errors:\n" + strings.Join(errs, "\n"))
 	}
 	return errors.Trace(verifyError)
-}
-
-// URLResolver is the part of charmrepo.Charmstore that we need to
-// resolve a charm url.
-type URLResolver interface {
-	ResolveWithPreferredChannel(*charm.URL, csparams.Channel) (*charm.URL, csparams.Channel, []string, error)
-}
-
-// ResolveBundleURL tries to interpret maybeBundle as a eharmstorr
-// bundle. If it turns out to be a bundle, the resolved URL and
-// channel are returned. If it isn't but there wasn't a problem
-// checking it, it returns a nil charm URL.
-func ResolveBundleURL(cstore URLResolver, maybeBundle string, preferredChannel csparams.Channel) (*charm.URL, csparams.Channel, error) {
-	userRequestedURL, err := charm.ParseURL(maybeBundle)
-	if err != nil {
-		return nil, "", errors.Trace(err)
-	}
-
-	// Charm or bundle has been supplied as a URL so we resolve and
-	// deploy using the store.
-	storeCharmOrBundleURL, channel, _, err := utils.ResolveCharm(cstore.ResolveWithPreferredChannel, userRequestedURL, preferredChannel)
-	if err != nil {
-		return nil, "", errors.Trace(err)
-	}
-	if storeCharmOrBundleURL.Series != "bundle" {
-		logger.Debugf(
-			`cannot interpret as charmstore bundle: %v (series) != "bundle"`,
-			storeCharmOrBundleURL.Series,
-		)
-		return nil, "", errors.NotValidf("charmstore bundle %q", maybeBundle)
-	}
-	return storeCharmOrBundleURL, channel, nil
-}
-
-// ResolvedBundle decorates a charm.Bundle instance with a type that implements
-// the charm.BundleDataSource interface.
-type ResolvedBundle struct {
-	parts []*charm.BundleDataPart
-}
-
-func NewResolvedBundle(b charm.Bundle) ResolvedBundle {
-	return ResolvedBundle{
-		parts: []*charm.BundleDataPart{
-			{
-				Data:        b.Data(),
-				PresenceMap: make(charm.FieldPresenceMap),
-			},
-		},
-	}
-}
-
-// Parts implements charm.BundleDataSource.
-func (rb ResolvedBundle) Parts() []*charm.BundleDataPart {
-	return rb.parts
-}
-
-// BasePath implements charm.BundleDataSource.
-func (ResolvedBundle) BasePath() string {
-	return ""
-}
-
-// ResolveInclude implements charm.BundleDataSource.
-func (ResolvedBundle) ResolveInclude(_ string) ([]byte, error) {
-	return nil, errors.NotSupportedf("remote bundle includes")
 }
