+++ cmd/juju/application/bundlediff.go
@@ -24,7 +24,7 @@ import (
 	"github.com/juju/juju/apiserver/params"
 	jujucmd "github.com/juju/juju/cmd"
 	appbundle "github.com/juju/juju/cmd/juju/application/bundle"
-	"github.com/juju/juju/cmd/juju/application/utils"
+	"github.com/juju/juju/cmd/juju/application/store"
 	"github.com/juju/juju/cmd/modelcmd"
 	"github.com/juju/juju/core/constraints"
 )
@@ -222,7 +222,7 @@ func (c *bundleDiffCommand) bundleDataSource(ctx *cmd.Context) (charm.BundleData
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
-	bundleURL, _, err := appbundle.ResolveBundleURL(charmStore, c.bundle, c.channel)
+	bundleURL, _, err := store.ResolveBundleURL(charmStore, c.bundle, c.channel)
 	if err != nil && !errors.IsNotValid(err) {
 		return nil, errors.Trace(err)
 	}
@@ -240,7 +240,7 @@ func (c *bundleDiffCommand) bundleDataSource(ctx *cmd.Context) (charm.BundleData
 		return nil, errors.Trace(err)
 	}
 
-	return appbundle.NewResolvedBundle(bundle), nil
+	return store.NewResolvedBundle(bundle), nil
 }
 
 func (c *bundleDiffCommand) charmStore() (BundleResolver, error) {
@@ -260,7 +260,7 @@ func (c *bundleDiffCommand) charmStore() (BundleResolver, error) {
 	if err != nil {
 		return nil, errors.Trace(err)
 	}
-	cstoreClient := utils.NewCharmStoreClient(bakeryClient, csURL).WithChannel(c.channel)
+	cstoreClient := store.NewCharmStoreClient(bakeryClient, csURL).WithChannel(c.channel)
 	return charmrepo.NewCharmStoreFromClient(cstoreClient), nil
 }
 
@@ -322,6 +322,6 @@ func (e *extractorImpl) Sequences() (map[string]int, error) {
 // BundleResolver defines what we need from a charm store to resolve a
 // bundle and read the bundle data.
 type BundleResolver interface {
-	utils.URLResolver
+	store.URLResolver
 	GetBundle(*charm.URL, string) (charm.Bundle, error)
 }
